make separate utils folder to declutter root folder
