+++ src/acp-client.ts
@@ -3,10 +3,10 @@ import { Readable, Writable } from "node:stream";
 import * as fs from "node:fs/promises";
 import * as acp from "@agentclientprotocol/sdk";
 import { executeCommand, killSession, type ExecutorSession } from "./executor.js";
-import { computeDiff } from "./diff.js";
-import { FileWatcher } from "./file-watcher.js";
+import { computeDiff } from "./utils/diff.js";
+import { FileWatcher } from "./utils/file-watcher.js";
 import * as path from "node:path";
-import { stripAnsi } from "./ansi.js";
+import { stripAnsi } from "./utils/ansi.js";
 import type { EventBus } from "./event-bus.js";
 import type { ContextManager } from "./context-manager.js";
 import type { Shell } from "./shell.js";
